chore: move categories
